# docs/swagger.yaml
basePath: /
definitions:
...
  entity.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  entity.LoginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
  entity.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  entity.UserInfo:
    properties:
      agent_id:
        type: string
      create_date:
        type: string
      first_number:
        type: string
      image:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      service_name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server Voice Transcribe server.
  title: Voice Transcribe API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/one:
    get:
      consumes:
      - application/json
      description: Get a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"